Move last test from sicknotes to the IT test
